requirements.txt
pydantic
flask
flask_cors
python-dotenv
bcrypt
pymongo
flask[async]
gunicorn
pydantic[email]